Refactored profile folder into profiles
