--- backend/internal/handlers/profile/routes.go
@@ -1,21 +1,0 @@
-package profile
-
-import (
-	"github.com/GenerateNU/nightlife/internal/auth"
-	"github.com/GenerateNU/nightlife/internal/types"
-	"github.com/gofiber/fiber/v2"
-)
-
-// Create Profile fiber route group (no unprotected routes)
-func Routes(app *fiber.App, params types.Params) {
-	service := newService(params.Store)
-
-	// Create Protected Grouping
-	protected := app.Group("/profile")
-
-	// Register Middleware
-	protected.Use(auth.Protected(&params.Supabase))
-
-	//Endpoints
-	protected.Post("/preferences", service.CreatePreferences)
-}
