--- backend/internal/handlers/profile/service.go
@@ -1,11 +1,0 @@
-package profile
-
-import "github.com/GenerateNU/nightlife/internal/storage"
-
-type Service struct {
-	store storage.Storage
-}
-
-func newService(store storage.Storage) *Service {
-	return &Service{store: store}
-}
