+++ backend/internal/handlers/profiles/profiles.go
@@ -10,6 +10,26 @@ import (
 	"github.com/google/uuid"
 )
 
+// POST Endpoint -> allows users to add their preferences to the db
+func (s *Service) CreatePreferences(c *fiber.Ctx) error {
+	var p models.Preferences
+	if err := c.BodyParser(&p); err != nil {
+		return errs.BadRequest(err)
+	}
+
+	if verrs := p.Validate(); verrs != nil {
+		return errs.InvalidRequestData(verrs)
+	}
+
+	if err := s.store.CreatePreferences(c.Context(), p); err != nil {
+		return err
+	}
+
+	// close out with success status
+	return c.Status(fiber.StatusCreated).JSON(p)
+
+}
+
 func (s *Service) UpdateProfilePreferences(c *fiber.Ctx) error {
 
 	// Parse the request body
